chore: move cursors demo to demos and add test for it
