+++ demos/cursors/cursors_test.go
@@ -23,7 +23,14 @@ import (
 	"github.com/gdamore/tcell/v3/vt"
 )
 
-// TestDemo tests the clipboard demo.
+func verifyCursor(t *testing.T, term vt.MockTerm, expected vt.CursorStyle) {
+	t.Helper()
+	if actual := term.Backend().GetCursor(); actual != expected {
+		t.Errorf("wrong cursor style %x != %x", actual, expected)
+	}
+}
+
+// TestDemo tests the cursors demo.
 func TestDemo(t *testing.T) {
 
 	mt := vt.NewMockTerm()
@@ -44,18 +51,37 @@ func TestDemo(t *testing.T) {
 	// this is needed because main is running asynchronously.
 	time.Sleep(time.Millisecond * 50)
 
-	mt.KeyEvent(vt.KeyEvent{Code: '1', Base: '1', Down: true})
-	mt.Drain()
+	mt.KeyEvent(vt.KeyEvent{Code: '1', Down: true})
+	time.Sleep(time.Millisecond * 20)
+	verifyCursor(t, mt, vt.BlinkingBlock)
 
-	mt.KeyEvent(vt.KeyEvent{Code: '2', Base: '2', Down: true})
-	mt.Drain()
+	mt.KeyEvent(vt.KeyEvent{Code: '2', Down: true})
+	time.Sleep(time.Millisecond * 20)
+	verifyCursor(t, mt, vt.SteadyBlock)
 
+	mt.KeyEvent(vt.KeyEvent{Code: '3', Down: true})
+	time.Sleep(time.Millisecond * 20)
+	verifyCursor(t, mt, vt.BlinkingUnderline)
+
+	mt.KeyEvent(vt.KeyEvent{Code: '4', Base: '4', Down: true})
+	time.Sleep(time.Millisecond * 20)
+	verifyCursor(t, mt, vt.SteadyUnderline)
+
+	mt.KeyEvent(vt.KeyEvent{Code: '5', Base: '5', Down: true})
+	time.Sleep(time.Millisecond * 20)
+	verifyCursor(t, mt, vt.BlinkingBar)
+
+	mt.KeyEvent(vt.KeyEvent{Code: '6', Base: '6', Down: true})
+	time.Sleep(time.Millisecond * 20)
+	verifyCursor(t, mt, vt.SteadyBar)
+
+	mt.KeyEvent(vt.KeyEvent{Code: '0', Base: '0', Down: true})
+	time.Sleep(time.Millisecond * 20)
+	verifyCursor(t, mt, vt.BlinkingBlock)
+
+	mt.KeyEvent(vt.KeyEvent{Code: 'L', Base: 'L', Mod: vt.ModCtrl, Down: true})
+	mt.Drain()
 	mt.KeyEvent(vt.KeyEvent{Code: 'Q', Base: 'Q', Mod: vt.ModCtrl, Down: true})
 	mt.Backend().GetSize()
 	wg.Wait()
-
-	expect := "Enjoy your new clipboard content!"
-	if result := string(mt.Backend().GetClipboard()); result != expect {
-		t.Errorf("clipboard content did not match! %q != %q", result, expect)
-	}
 }
